refactoring instance lifecycle to the proper class
