+++ packages/b2c-tooling-sdk/src/cli/cartridge-command.ts
@@ -8,16 +8,6 @@ import {Args, Command, Flags} from '@oclif/core';
 import {InstanceCommand} from './instance-command.js';
 import {findCartridges, type CartridgeMapping, type FindCartridgesOptions} from '../operations/code/cartridges.js';
 import {
-  B2CLifecycleRunner,
-  createB2COperationContext,
-  type B2COperationType,
-  type B2COperationContext,
-  type B2COperationResult,
-  type BeforeB2COperationResult,
-  type B2COperationLifecycleHookOptions,
-  type B2COperationLifecycleHookResult,
-} from './lifecycle.js';
-import {
   CartridgeProviderRunner,
   type CartridgeDiscoveryOptions,
   type CartridgeProvidersHookOptions,
@@ -64,46 +54,15 @@ export abstract class CartridgeCommand<T extends typeof Command> extends Instanc
     }),
   };
 
-  /** Lifecycle runner for B2C operation hooks */
-  protected lifecycleRunner?: B2CLifecycleRunner;
-
   /** Cartridge provider runner for custom cartridge discovery */
   protected cartridgeProviderRunner?: CartridgeProviderRunner;
 
   /**
-   * Override init to collect lifecycle providers and cartridge providers from plugins.
+   * Override init to collect cartridge providers from plugins.
    */
   public async init(): Promise<void> {
     await super.init();
-    await this.collectLifecycleProviders();
     await this.collectCartridgeProviders();
-  }
-
-  /**
-   * Collects lifecycle providers from plugins via the `b2c:operation-lifecycle` hook.
-   */
-  protected async collectLifecycleProviders(): Promise<void> {
-    this.lifecycleRunner = new B2CLifecycleRunner(this.logger);
-
-    const hookOptions: B2COperationLifecycleHookOptions = {
-      flags: this.flags as Record<string, unknown>,
-    };
-
-    const hookResult = await this.config.runHook('b2c:operation-lifecycle', hookOptions);
-
-    for (const success of hookResult.successes) {
-      const result = success.result as B2COperationLifecycleHookResult | undefined;
-      if (!result?.providers?.length) continue;
-      this.lifecycleRunner.addProviders(result.providers);
-    }
-
-    for (const failure of hookResult.failures) {
-      this.logger?.warn(`Plugin ${failure.plugin.name} b2c:operation-lifecycle hook failed: ${failure.error.message}`);
-    }
-
-    if (this.lifecycleRunner.size > 0) {
-      this.logger?.debug(`Registered ${this.lifecycleRunner.size} lifecycle provider(s)`);
-    }
   }
 
   /**
@@ -138,33 +97,6 @@ export abstract class CartridgeCommand<T extends typeof Command> extends Instanc
     if (providerCount > 0 || transformerCount > 0) {
       this.logger?.debug(`Registered ${providerCount} cartridge provider(s) and ${transformerCount} transformer(s)`);
     }
-  }
-
-  /**
-   * Creates a B2C operation context for lifecycle hooks.
-   */
-  protected createContext(operationType: B2COperationType, metadata: Record<string, unknown>): B2COperationContext {
-    return createB2COperationContext(operationType, metadata, this.instance);
-  }
-
-  /**
-   * Runs beforeOperation hooks for all providers.
-   */
-  protected async runBeforeHooks(context: B2COperationContext): Promise<BeforeB2COperationResult> {
-    if (!this.lifecycleRunner) {
-      return {};
-    }
-    return this.lifecycleRunner.runBefore(context);
-  }
-
-  /**
-   * Runs afterOperation hooks for all providers.
-   */
-  protected async runAfterHooks(context: B2COperationContext, result: B2COperationResult): Promise<void> {
-    if (!this.lifecycleRunner) {
-      return;
-    }
-    await this.lifecycleRunner.runAfter(context, result);
   }
 
   /**
